faq-system-master/requirements.txt
Django==5.1.5
djangorestframework==3.14.0
django-ckeditor==6.7.2
redis==5.0.1
googletrans==3.1.0a0
django-cors-headers==4.3.1
django-js-asset==3.0.1
python-dotenv==1.0.1
deep-translator==1.11.4
flake8==7.0.0
coverage==7.4.1
pytest==8.0.0
pytest-django==4.8.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
